rename bokeh.py to bk.py, otherwise there's an import issue in _viz_bokeh.py
